changed project structure
