--- Player.cpp
@@ -1,7 +1,0 @@
-//
-// Created by georgiana on 24.11.2020.
-//
-
-
-#include "headers/Player.h"
-
