+++ classes/Enemy.cpp
@@ -4,13 +4,11 @@
 
 #include "headers/Enemy.h"
 
-//todo niste enum-uri sau ceva pentru diferite clase de inamici ca sa fie mai usor de creat? sau cu citire din fisier
-///constructor of enemy class with specific atributes
-Enemy::Enemy(int h, int s, std::string sprint,QWidget * parent) {
+//todo niste enum-uri sau ceva pentru diferite clase de inamici ca sa fie mai usor de creat?
+
+Enemy::Enemy(int h, int s) {
     this->health=h;
     this->speed=s;
-    sprintName=sprint;
-    this->setParent(parent);
 }
 
 ///level 1 -> general type of enemy
@@ -39,32 +37,15 @@ void Enemy::die() {
     //alive=false
 }
 
-int Enemy::getX() {
-    return x;
-}
+/*
+ * // vine waveul
+ * int nrinamici = wave[0].length;
+ *
+ * if(enemy.health == 0)
+ * enemy.die();
+ * nrinamici--;
+ * 
+ */
 
-int Enemy::getY() {
-    return y;
-}
-
-///move the enemy to the (nx,ny) coordinates
-//void Enemy::move(int nx, int ny) {
-//    x=nx;
-//    y=ny;
-//    ///todo further tests will have to be implemented
-//}
-
-void Enemy::setCoordinates(int x, int y) {
-    this->x=x;
-    this->y=y;
-}
-
-void Enemy::reposition() {
-    this->setGeometry(QRect(300,200,100,100));
-}
-
-//void Enemy::reposition(QMouseEvent *event) {
-//    this->setGeometry(QRect(0,0,event->pos().x(),event->pos().y()));
-//}
