+++ x64/Debug/PIU_Project.log
@@ -1,2 +1,3 @@
-﻿  StartGame.cpp
+﻿  LINK : C:\Users\Ionut\Desktop\FACULTATE\FACULTATE\Anul4\Sem1\PIU\PIU_Git\PIU-project\x64\Debug\PIU_Project.exe not found or not built by the last incremental link; performing full link
+     Creating library C:\Users\Ionut\Desktop\FACULTATE\FACULTATE\Anul4\Sem1\PIU\PIU_Git\PIU-project\x64\Debug\PIU_Project.lib and object C:\Users\Ionut\Desktop\FACULTATE\FACULTATE\Anul4\Sem1\PIU\PIU_Git\PIU-project\x64\Debug\PIU_Project.exp
   PIU_Project.vcxproj -> C:\Users\Ionut\Desktop\FACULTATE\FACULTATE\Anul4\Sem1\PIU\PIU_Git\PIU-project\x64\Debug\PIU_Project.exe
