+++ x64/Debug/qmake/temp/3D7FC4E1-974B-4DD6-B274-B48EBB517275.PIU_Project.designtime.idx
@@ -1,1 +1,1 @@
-C:\Users\Ionut\AppData\Local\Temp\hbxo1ih5.lbv.designtime.props
+C:\Users\Ionut\AppData\Local\Temp\iwswk3bm.1tx.designtime.props
