+++ x64/Release/qmake/temp/3D7FC4E1-974B-4DD6-B274-B48EBB517275.PIU_Project.designtime.idx
@@ -1,1 +1,1 @@
-C:\Users\Ionut\AppData\Local\Temp\rsdn5olp.5fu.designtime.props
+C:\Users\Ionut\AppData\Local\Temp\xza1he4x.eeg.designtime.props
